+++ idp_engine/Expression.py
@@ -565,6 +565,15 @@ class Subtype(Symbol):
         return self.name + ("" if not self.out else
                             f"[{'*'.join(str(s) for s in self.ins)}->{self.out}]")
 
+    def __eq__(self, other):
+        self.check(self.name != CONCEPT or self.out,
+                   f"`Concept` must be qualified with a type signature")
+        return (self.name == other.name and
+                (not self.out or (
+                    self.out == other.out and
+                    len(self.ins) == len(other.ins) and
+                    all(s==o for s, o in zip(self.ins, other.ins)))))
+
     def range():
         pass  # monkey-patched