+++ idp_engine/Simplify.py
@@ -35,7 +35,7 @@ from .Expression import (
     Number, Date, Brackets, TRUE, FALSE, NOT, AND, OR)
 from .Parse import Symbol, Enumeration, Tuple, TypeDeclaration
 from .Assignments import Status as S, Assignment
-from .utils import BOOL, INT, CONCEPT, ABS, ARITY, INPUT_DOMAIN, OUTPUT_DOMAIN, RESERVED_SYMBOLS
+from .utils import BOOL, INT, CONCEPT, ABS, RESERVED_SYMBOLS
 
 
 # class Expression  ###########################################################
@@ -256,7 +256,8 @@ def update_arith(self, family, operands):
                 f"Incorrect numeric type in {self}")
         out = operands1[0].py_value
 
-        for e, op in zip(operands1[1:], self.operator):
+        assert len(self.operator) == len(operands1[1:]), "Internal error"
+        for op, e in zip(self.operator, operands1[1:]):
             function = Operator.MAP[op]
 
             if op == '/' and self.type == INT:  # integer division
@@ -376,42 +377,6 @@ def update_exprs(self, new_exprs):
         if self.decl.name == new_exprs[0].decl.tester.name:
             return self._change(value=TRUE, sub_exprs=new_exprs)
 
-    # simplify arity(s), input_domain(s,0), output_domain(s)
-    if (self.decl and new_exprs
-        and type(new_exprs[0]) == UnappliedSymbol and new_exprs[0].decl):
-        if self.decl.name == ARITY:
-            self.check(len(new_exprs) == 1,
-                    f"Incorrect number of arguments for '{ARITY}': {len(new_exprs)}")
-            self.check(new_exprs[0].decl.type == CONCEPT,
-                    f"Argument of '{ARITY}' must be a Concept: {new_exprs[0]}")
-            value = Number(number=str(new_exprs[0].decl.symbol.decl.arity))
-            return self._change(value=value, sub_exprs=new_exprs)
-        elif self.decl.name == INPUT_DOMAIN:
-            self.check(len(new_exprs) == 2,
-                    f"Incorrect number of arguments for '{INPUT_DOMAIN}': {len(new_exprs)}")
-            self.check(new_exprs[0].decl.type == CONCEPT
-                       or (new_exprs[0].decl.sort.decl.arity == 1
-                           and new_exprs[0].decl.sort.decl.type == BOOL),
-                    f"First argument of '{INPUT_DOMAIN}' must be a Concept: {new_exprs[0]}")
-            self.check(new_exprs[1].type == INT,
-                    f"Second argument of '{INPUT_DOMAIN}' must be a Int: {new_exprs[1]}")
-            if isinstance(new_exprs[1], Number):
-                # find the Concept for the input domain
-                symbol_string = f"`{new_exprs[0].decl.symbol.decl.sorts[new_exprs[1].py_value - 1]}"
-                value = self.decl.out.decl.map[symbol_string]
-                return self._change(value=value, sub_exprs=new_exprs)
-        elif self.decl.name == OUTPUT_DOMAIN:
-            self.check(len(new_exprs) == 1,
-                    f"Incorrect number of arguments for '{OUTPUT_DOMAIN}': {len(new_exprs)}")
-            self.check(new_exprs[0].decl.type == CONCEPT,
-                    f"Argument of '{OUTPUT_DOMAIN}' must be a Concept: {new_exprs[0]}")
-            # find the Concept for the output domain of the argument
-            symbol_string = f"`{new_exprs[0].decl.symbol.decl.out}"
-            value = self.decl.out.decl.map[symbol_string]
-            return self._change(value=value, sub_exprs=new_exprs)
-        elif type(self.decl) == TypeDeclaration:  # type
-            value = TRUE if self.sub_exprs[0] in self.decl.range else FALSE
-            return self._change(value=value)
     return self._change(sub_exprs=new_exprs)
 AppliedSymbol.update_exprs = update_exprs